+++ NEWS.md
@@ -28,6 +28,8 @@ New features
   - New configuration ``dice_numeral`` which provide six-sided dice with numerals instead of pips in six color schemes (#311).
   - New configuration ``dice_d8`` which provide eight-sided dice in six color schemes (color controlled by their suit).
     Octahedrons with the rank as a numeral at the top face (#312).
+  - New configuration ``dice_d12`` which provide twelve-sided dice in six color schemes (color controlled by their suit).
+    Dodecahedrons with the rank as a numeral at the top face (#316).
   - New configuration ``dice_d20`` which provide twenty-sided dice in six color schemes (color controlled by their suit).
     Icosahedrons with the rank as a numeral at the top face (#317).
   - New configurations ``dominoes_chinese`` and ``dominoes_chinese_black`` which provide [Chinese dominoes](https://en.wikipedia.org/wiki/Chinese_dominoes) and [Asian-style dice](https://en.wikipedia.org/wiki/Dice#Arrangement).